Splitting Linux specific parts into a separate file
