+++ pkg/pillar/containerd/containerd.go
@@ -16,7 +16,6 @@ import (
 	"strings"
 	"time"
 
-	"golang.org/x/sys/unix"
 	"google.golang.org/grpc/connectivity"
 
 	"github.com/containerd/containerd"
@@ -778,29 +777,6 @@ func (client *Client) verifyCtr(ctx context.Context, verifyCtx bool) error {
 		if ctx.Err() == context.Canceled {
 			return fmt.Errorf("verifyCtr: Containerd context is calcelled")
 		}
-	}
-	return nil
-}
-
-// bind mount a namespace file
-func bindNS(ns string, path string, pid int) error {
-	if path == "" {
-		return nil
-	}
-	// the path and file need to exist for the bind to succeed, so try to create
-	dir := filepath.Dir(path)
-	if err := os.MkdirAll(dir, 0755); err != nil {
-		return fmt.Errorf("bindNS: Cannot create leading directories %s for bind mount destination: %v", dir, err)
-	}
-	fi, err := os.Create(path)
-	if err != nil {
-		return fmt.Errorf("bindNS: Cannot create a mount point for namespace bind at %s: %v", path, err)
-	}
-	if err := fi.Close(); err != nil {
-		return err
-	}
-	if err := unix.Mount(fmt.Sprintf("/proc/%d/ns/%s", pid, ns), path, "", unix.MS_BIND, ""); err != nil {
-		return fmt.Errorf("bindNS: Failed to bind %s namespace at %s: %v", ns, path, err)
 	}
 	return nil
 }
